#9 rename from interface to contract
