+++ contract/console/kernel.go
