+++ contract/decorator/register_service_provider.go
