+++ contract/http/kernel.go
