+++ src/routes/api.go
@@ -16,7 +16,8 @@ import (
 |
 */
 var Api = []routing.RouteCollection{
-	routing.Get("/users/{ids}", controllers.User.Index),
+	routing.Get("/users", controllers.User.Index),
+	routing.Get("/users/{ids}", controllers.User.Show),
 	routing.Post("/users", controllers.User.Store),
 	routing.Delete("/users/{ids?}", controllers.User.Destroy),
 }
